backend/pyproject.toml
[project]
name = "claude-hub"
version = "0.1.0"
description = "Claude Hub — Kanban board orchestrating Claude Code CLI sessions"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.84.0",
    "fastapi[standard]>=0.135.1",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "redis[hiredis]>=7.3.0",
    "websockets>=16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_hub"]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
